Move middleware into one file per app for simplicity
